Add a Systemd user service file for Steam Deck
The Steam Deck has Systemd user instances enabled.  This file has all
the information needed for steam-presence to be launched automatically,
under Systemd, as the Steam Deck's `deck` user, when the Steam Deck
boots.

This Systemd service file has a number of assumptions:

* It assumes that steam-presence is at path `/home/deck/steam-presence`.

* It assumes that a Python venv has been created in
  `/home/deck/steam-presence`.

It's easy to change those assumptions, by changing the `ExecStart` line
of the file.

There are also a **lot** of comments, both explaining the parts of the
Systemd service file, and providing users instructions on how to prepare
steam-presence on their Steam Deck.